api/requirements.txt
flask
flask-sqlalchemy
flask-login
psycopg2-binary
flask-sock
python-dotenv
vonage
websocket-client
pywebpush
gunicorn
werkzeug
gevent
google-genai
flask-cors
